+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `bonsaidb::client::Error::Core`'s `Display` no longer just prints "unexpected
   disconnection". Instead, the inner error's `Display` is now displayed.
+- `bonsaidb::client::Async`/`Blocking` are now exposed. These types are used
+  when building a client.
+- `bonsaidb::server::Backend::client_authenticated` is now invoked.
 
 ## v0.5.0